+++ scripts/test/Selenium/Output/Ticket/OverviewSmall.t
@@ -37,8 +37,10 @@ $Selenium->RunTest(
             UserLogin => $TestUserLogin,
         );
 
+        my $RandomID = $Helper->GetRandomID();
+
         # create test queue
-        my $QueueName = 'Queue' . $Helper->GetRandomID();
+        my $QueueName = 'Queue' . $RandomID;
         my $QueueID   = $Kernel::OM->Get('Kernel::System::Queue')->QueueAdd(
             Name            => $QueueName,
             ValidID         => 1,
@@ -54,6 +56,32 @@ $Selenium->RunTest(
             "QueueAdd() successful for test $QueueName - ID $QueueID",
         );
 
+        # get user object
+        my $UserObject = $Kernel::OM->Get('Kernel::System::User');
+
+        # create test users
+        my @UserIDs;
+        my %Users;
+        for my $User ( 1 .. 15 ) {
+            my $UserFirstname = 'Firstname' . $User;
+            my $UserLastname  = 'Lastname' . $User;
+            my $UserID        = $UserObject->UserAdd(
+                UserFirstname => $UserFirstname,
+                UserLastname  => $UserLastname,
+                UserLogin     => 'test' . $RandomID . $User,
+                UserEmail     => "test$RandomID$User\@example.com",
+                ValidID       => 1,
+                ChangeUserID  => $TestUserID,
+            );
+            $Self->True(
+                $UserID,
+                "User is created - $UserID",
+            );
+
+            push @UserIDs, $UserID;
+            $Users{$UserID} = "$UserFirstname $UserLastname";
+        }
+
         # get ticket object
         my $TicketObject = $Kernel::OM->Get('Kernel::System::Ticket');
 
@@ -63,16 +91,17 @@ $Selenium->RunTest(
         for my $Ticket ( 1 .. 15 ) {
             my $TicketNumber = $TicketObject->TicketCreateNumber();
             my $TicketID     = $TicketObject->TicketCreate(
-                TN           => $TicketNumber,
-                Title        => 'Some Ticket Title',
-                Queue        => $QueueName,
-                Lock         => 'unlock',
-                Priority     => '3 normal',
-                State        => 'new',
-                CustomerID   => 'TestCustomer',
-                CustomerUser => 'customer@example.com',
-                OwnerID      => $TestUserID,
-                UserID       => $TestUserID,
+                TN            => $TicketNumber,
+                Title         => 'Some Ticket Title',
+                Queue         => $QueueName,
+                Lock          => 'unlock',
+                Priority      => '3 normal',
+                State         => 'new',
+                CustomerID    => 'TestCustomer',
+                CustomerUser  => 'customer@example.com',
+                OwnerID       => $UserIDs[ $Ticket - 1 ],
+                ResponsibleID => $UserIDs[ $Ticket - 1 ],
+                UserID        => $TestUserID,
             );
 
             $Self->True(
@@ -109,6 +138,21 @@ $Selenium->RunTest(
         $Selenium->execute_script(
             "\$('#UserTicketOverviewSmallPageShown').val('10').trigger('redraw.InputField').trigger('change');"
         );
+
+        # move responsible from left to the right side
+        $Selenium->mouse_move_to_location(
+            element => $Selenium->find_element( '//li[@data-fieldname="Responsible"]', 'xpath' ),
+        );
+
+        $Selenium->DragAndDrop(
+            Element      => 'li[data-fieldname="Responsible"]',
+            Target       => '#AssignedFields-DashboardAgentTicketStatusView',
+            TargetOffset => {
+                X => 185,
+                Y => 10,
+            },
+        );
+
         $Selenium->find_element( "#DialogButton1", 'css' )->VerifiedClick();
 
         # sort by ticket number, order up
@@ -141,6 +185,56 @@ $Selenium->RunTest(
             "$SortTicketNumbers[14] - found on screen"
         );
 
+        # check if owner and responsible columns are sorted by names (bug#4439)
+        for my $Column (qw(Responsible Owner)) {
+
+            # sort by column, order up
+            $Selenium->find_element("//a[\@name='OverviewControl'][contains(\@title, '$Column')]")->VerifiedClick();
+
+            # check for user with 1 in their name on first page
+            # and verify that user with 9 in their name is not present
+            $Self->True(
+                index( $Selenium->get_page_source(), $Users{ $UserIDs[0] } ) > -1,
+                "$Users{ $UserIDs[0] } - found on screen"
+            );
+            $Self->True(
+                index( $Selenium->get_page_source(), $Users{ $UserIDs[8] } ) == -1,
+                "$Users{ $UserIDs[8] } - not found on screen"
+            );
+
+            # switch to 2nd page to test pagination
+            $Selenium->find_element( "#AgentTicketStatusViewPage2", 'css' )->VerifiedClick();
+
+            # check for user with 9 in their name
+            $Self->True(
+                index( $Selenium->get_page_source(), $Users{ $UserIDs[8] } ) > -1,
+                "$Users{ $UserIDs[8] } - found on screen"
+            );
+
+            # sort by column, order down
+            $Selenium->find_element("//a[\@name='OverviewControl'][contains(\@title, '$Column')]")->VerifiedClick();
+
+            # check for user with 9 in their name on first page
+            # and verify that user with 1 in their name is not present
+            $Self->True(
+                index( $Selenium->get_page_source(), $Users{ $UserIDs[8] } ) > -1,
+                "$Users{ $UserIDs[8] } - found on screen"
+            );
+            $Self->True(
+                index( $Selenium->get_page_source(), $Users{ $UserIDs[0] } ) == -1,
+                "$Users{ $UserIDs[0] } - not found on screen"
+            );
+
+            # switch to 2nd page to test pagination
+            $Selenium->find_element( "#AgentTicketStatusViewPage2", 'css' )->VerifiedClick();
+
+            # check for user with 1 in their name
+            $Self->True(
+                index( $Selenium->get_page_source(), $Users{ $UserIDs[0] } ) > -1,
+                "$Users{ $UserIDs[0] } - found on screen"
+            );
+        }
+
         # remove all filters
         $Selenium->find_element( "li.ContextSettings.RemoveFilters a", 'css' )->VerifiedClick();
 
@@ -157,8 +251,28 @@ $Selenium->RunTest(
             );
         }
 
+        my $DBObject = $Kernel::OM->Get('Kernel::System::DB');
+
+        # delete created test users
+        for my $UserID (@UserIDs) {
+            $Success = $DBObject->Do(
+                SQL => "DELETE FROM user_preferences WHERE user_id = $UserID",
+            );
+            $Self->True(
+                $Success,
+                "Delete user preferences - $UserID",
+            );
+            $Success = $DBObject->Do(
+                SQL => "DELETE FROM users WHERE id = $UserID",
+            );
+            $Self->True(
+                $Success,
+                "Delete user - $UserID",
+            );
+        }
+
         # delete created test queue
-        $Success = $Kernel::OM->Get('Kernel::System::DB')->Do(
+        $Success = $DBObject->Do(
             SQL => "DELETE FROM queue WHERE id = $QueueID",
         );
         $Self->True(
@@ -167,12 +281,12 @@ $Selenium->RunTest(
         );
 
         # make sure cache is correct
-        for my $Cache (qw( Ticket Queue )) {
+        for my $Cache (qw( Ticket Queue User )) {
             $Kernel::OM->Get('Kernel::System::Cache')->CleanUp(
                 Type => $Cache,
             );
         }
-    }
+    },
 );
 
 1;